# Improve performance and reduce allocations for request and binding parameters