--- a/src/Ninject/Planning/Bindings/Binding.cs
+++ b/src/Ninject/Planning/Bindings/Binding.cs
@@ -175,7 +175,7 @@ namespace Ninject.Planning.Bindings
         /// <summary>
         /// Gets the parameters defined for the binding.
         /// </summary>
-        public ICollection<IParameter> Parameters
+        public IList<IParameter> Parameters
         {
             get
             {