--- a/src/Ninject/ReadOnlyKernel.cs
+++ b/src/Ninject/ReadOnlyKernel.cs
@@ -182,7 +182,7 @@ namespace Ninject
         /// </returns>
         /// <exception cref="ArgumentNullException"><paramref name="service"/> is <see langword="null"/>.</exception>
         /// <exception cref="ArgumentNullException"><paramref name="parameters"/> is <see langword="null"/>.</exception>
-        public IRequest CreateRequest(Type service, Func<IBindingMetadata, bool> constraint, IEnumerable<IParameter> parameters, bool isOptional, bool isUnique)
+        public IRequest CreateRequest(Type service, Func<IBindingMetadata, bool> constraint, IReadOnlyList<IParameter> parameters, bool isOptional, bool isUnique)
         {
             return new Request(service, constraint, parameters, null, isOptional, isUnique);
         }
@@ -326,7 +326,7 @@ namespace Ninject
             {
                 if (request.ParentRequest == null)
                 {
-                    request = this.CreateRequest(service, null, request.Parameters.Where(p => p.ShouldInherit), true, false);
+                    request = this.CreateRequest(service, null, request.Parameters.GetShouldInheritParameters(), true, false);
                 }
                 else
                 {