… CHANGELOG.md
… src/Ninject/Syntax/ResolutionExtensions.cs